fix: fixed the bug
